+++ src/ksi/pkitruststore_openssl.c
@@ -826,7 +826,7 @@ static int pki_truststore_verifyCertificateConstraints(const KSI_PKITruststore *
 	for (i = 0; certConstraints[i].oid != NULL; i++) {
 		KSI_CertConstraint *ptr = &certConstraints[i];
 
-		KSI_LOG_info(pki->ctx, "%d. Verifying PKI signature certificate with OID: '%s' expected value: '%s'.", i + 1, ptr->oid, ptr->val);
+		KSI_LOG_info(pki->ctx, "%llu. Verifying PKI signature certificate with OID: '%s' expected value: '%s'.", (unsigned long long)i + 1, ptr->oid, ptr->val);
 
 		oid = OBJ_txt2obj(ptr->oid, 1);
 		if (oid == NULL) {
